+++ near-contract-standards/src/lib.rs
@@ -8,6 +8,4 @@ pub mod storage_management;
 /// blob, allowing it to be stored for a period of time before deployed.
 pub mod upgrade;
 
-/// Log JSON events to be caught by indexer
-pub mod event;
-pub use event::NearEvent;
+pub(crate) mod event;